pyproject.toml
[project]
name = "drone-motor-catalog-llm-parser"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "scrapegraphai>=1.13",
    "pandas>=1.3",
    "pydantic>=2.0",
    "pillow>=10.0",
    "ollama>=0.1",
]
